+++ CHANGELOG.md
@@ -8,6 +8,10 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
 
 ##### Enhancements
 
+* Simplify logic around framework input and output paths  
+  [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
+  [#6803](https://github.com/CocoaPods/CocoaPods/pull/6803)
+
 * Add inputs and outputs to check manifest lock and embed framework script phases  
   [Dimitris Koutsogiorgas](https://github.com/dnkoutso)
   [#6797](https://github.com/CocoaPods/CocoaPods/issues/6797)